requirements.txt
python-dotenv
Pillow
pdf2image
google-cloud-vision
google-generativeai
google-genai